+++ x-pack/legacy/plugins/siem/public/containers/detection_engine/signals/errors_types/privilege_user_error.ts
@@ -16,6 +16,7 @@ export class PrivilegeUserError extends Error {
     this.message = errObj.message ?? '';
     this.statusCode = errObj.statusCode ?? -1;
     this.error = errObj.error ?? '';
+    this.name = 'PrivilegeUserError';
 
     // Set the prototype explicitly.
     Object.setPrototypeOf(this, PrivilegeUserError.prototype);
